moving components folder, deleting comments
