+++ src/components/ArtistMainCard.js
