+++ src/pages/App.js
@@ -1,6 +1,6 @@
-import Header from "../assets/components/common/Header";
-import FootMain from "../assets/components/common/FootMain";
-import ScrollToTop from "../assets/components/common/ScrollToTop";
+import Header from "../components/common/Header";
+import FootMain from "../components/common/FootMain";
+import ScrollToTop from "../components/common/ScrollToTop";
 import { Outlet } from "react-router-dom";
 
 function App() {
